+++ bcbio/variation/varscan.py
@@ -5,22 +5,18 @@ http://varscan.sourceforge.net/
 
 from collections import namedtuple
 import contextlib
-from distutils.version import LooseVersion
 import os
-import shutil
+import sys
 
 from bcbio import broad, utils
 from bcbio.distributed.transaction import file_transaction, tx_tmpdir
 from bcbio.pipeline import config_utils
-from bcbio.provenance import do, programs
-from bcbio.utils import file_exists, append_stem
-from bcbio.variation import freebayes, samtools, vcfutils
+from bcbio.provenance import do
+from bcbio.variation import samtools, vcfutils
 from bcbio.variation.vcfutils import (combine_variant_files, write_empty_vcf,
-                                      get_paired_bams, is_paired_analysis,
-                                      bgzip_and_index, move_vcf)
+                                      get_paired_bams, bgzip_and_index)
 
 import pysam
-import vcf
 
 
 def run_varscan(align_bams, items, ref_file, assoc_files,
@@ -60,181 +56,77 @@ def _varscan_paired(align_bams, ref_file, items, target_regions, out_file):
 
     max_read_depth = "1000"
     config = items[0]["config"]
-
-    version = programs.jar_versioner("varscan", "VarScan")(config)
-    if LooseVersion(version) < LooseVersion("v2.3.6"):
-        raise IOError(
-            "Please install version 2.3.6 or better of VarScan with support "
-            "for multisample calling and indels in VCF format.")
-    varscan_jar = config_utils.get_jar(
-        "VarScan",
-        config_utils.get_program("varscan", config, "dir"))
-
-    remove_zerocoverage = "grep -v -P '\t0\t\t$'"
-
-    # No need for names in VarScan, hence the "_"
-
     paired = get_paired_bams(align_bams, items)
     if not paired.normal_bam:
         affected_batch = items[0]["metadata"]["batch"]
         message = ("Batch {} requires both tumor and normal BAM files for"
-            " VarScan cancer calling").format(affected_batch)
+                   " VarScan cancer calling").format(affected_batch)
         raise ValueError(message)
 
-    if not file_exists(out_file):
-        orig_out_file = out_file
-        out_file = orig_out_file.replace(".vcf.gz", ".vcf")
+    if not utils.file_exists(out_file):
+        assert out_file.endswith(".vcf.gz"), "Expect bgzipped output to VarScan"
+        normal_mpileup_cl = samtools.prep_mpileup([paired.normal_bam], ref_file,
+                                                  config, max_read_depth,
+                                                  target_regions=target_regions,
+                                                  want_bcf=False)
+        tumor_mpileup_cl = samtools.prep_mpileup([paired.tumor_bam], ref_file,
+                                                 config, max_read_depth,
+                                                 target_regions=target_regions,
+                                                 want_bcf=False)
         base, ext = utils.splitext_plus(out_file)
-        cleanup_files = []
-        for fname, mpext in [(paired.normal_bam, "normal"), (paired.tumor_bam, "tumor")]:
-            mpfile = "%s-%s.mpileup" % (base, mpext)
-            cleanup_files.append(mpfile)
-            with file_transaction(config, mpfile) as mpfile_tx:
-                mpileup = samtools.prep_mpileup([fname], ref_file,
-                                                config, max_read_depth,
-                                                target_regions=target_regions,
-                                                want_bcf=False)
-                cmd = "{mpileup} > {mpfile_tx}"
-                cmd = cmd.format(**locals())
-                do.run(cmd, "samtools mpileup".format(**locals()), None,
-                       [do.file_exists(mpfile_tx)])
-
-        # Sometimes mpileup writes an empty file: in this case we
-        # just skip the rest of the analysis (VarScan will hang otherwise)
-
-        if any(os.stat(filename).st_size == 0 for filename in cleanup_files):
-            write_empty_vcf(orig_out_file, config)
-            return
-
-        # First index is normal, second is tumor
-        normal_tmp_mpileup = cleanup_files[0]
-        tumor_tmp_mpileup = cleanup_files[1]
-
         indel_file = base + ".indel.vcf"
         snp_file = base + ".snp.vcf"
-        cleanup_files.append(indel_file)
-        cleanup_files.append(snp_file)
         with file_transaction(config, indel_file, snp_file) as (tx_indel, tx_snp):
             with tx_tmpdir(items[0]) as tmp_dir:
                 jvm_opts = _get_varscan_opts(config, tmp_dir)
-                fix_ambig = vcfutils.fix_ambiguous_cl()
-                tx_snp_in = "%s-orig" % os.path.splitext(tx_snp)[0]
-                tx_indel_in = "%s-orig" % os.path.splitext(tx_indel)[0]
-                varscan_cmd = ("java {jvm_opts} -jar {varscan_jar} somatic"
-                               " {normal_tmp_mpileup} {tumor_tmp_mpileup} "
-                               "--output-snp {tx_snp_in} --output-indel {tx_indel_in} "
+                remove_zerocoverage = "grep -v -P '\t0\t\t$'"
+                varscan_cmd = ("varscan {jvm_opts} somatic "
+                               " <({normal_mpileup_cl} | {remove_zerocoverage}) "
+                               "<({tumor_mpileup_cl} | {remove_zerocoverage}) "
+                               "--output-snp {tx_snp} --output-indel {tx_indel} "
                                " --output-vcf --min-coverage 5 --p-value 0.98 "
                                "--strand-filter 1 ")
                 # add minimum AF
                 if "--min-var-freq" not in varscan_cmd:
                     min_af = float(utils.get_in(paired.tumor_config, ("algorithm",
-                                                                      "min_allele_fraction"),10)) / 100.0
+                                                                      "min_allele_fraction"), 10)) / 100.0
                     varscan_cmd += "--min-var-freq {min_af} "
                 do.run(varscan_cmd.format(**locals()), "Varscan", None, None)
-                for orig_fname, fname in [(tx_snp_in, tx_snp), (tx_indel_in, tx_indel)]:
-                    cmd = "vcfuniqalleles {orig_fname}.vcf | {fix_ambig} > {fname}"
-                    do.run(cmd.format(**locals()), "Varscan paired fix")
 
-        # VarScan files need to be corrected to match the VCF specification
-        # We do this before combining them otherwise merging may fail
-        # if there are invalid records
         to_combine = []
-        if do.file_exists(snp_file):
-            to_combine.append(snp_file)
-            _fix_varscan_vcf(snp_file, paired.normal_name, paired.tumor_name, config)
-
-        if do.file_exists(indel_file):
-            to_combine.append(indel_file)
-            _fix_varscan_vcf(indel_file, paired.normal_name, paired.tumor_name, config)
+        for fname in [snp_file, indel_file]:
+            if utils.file_exists(fname):
+                fix_file = "%s-fix.vcf.gz" % (utils.splitext_plus(fname)[0])
+                with file_transaction(config, fix_file) as tx_fix_file:
+                    fix_ambig = vcfutils.fix_ambiguous_cl()
+                    py_cl = os.path.join(os.path.dirname(sys.executable), "py")
+                    normal_name = paired.normal_name
+                    tumor_name = paired.tumor_name
+                    cmd = ("cat {fname} | "
+                           "{py_cl} -x 'bcbio.variation.varscan.fix_varscan_output(x,"
+                            """ "{normal_name}", "{tumor_name}")' | """
+                           "{fix_ambig} | ifne vcfuniqalleles | "
+                           """bcftools filter -m + -s REJECT -e 'SS != "." && SS != "2"' 2> /dev/null | """
+                           " bgzip -c > {tx_fix_file}")
+                    do.run(cmd.format(**locals()), "Varscan paired fix")
+                to_combine.append(fix_file)
 
         if not to_combine:
-            write_empty_vcf(orig_out_file, config)
-            return
-
-        out_file = combine_variant_files([snp_file, indel_file],
-                                         out_file, ref_file, config,
-                                         region=target_regions)
-
-        # Remove cleanup files
-
-        for extra_file in cleanup_files:
-            for ext in ["", ".gz", ".gz.tbi"]:
-                if os.path.exists(extra_file + ext):
-                    os.remove(extra_file + ext)
-
+            out_file = write_empty_vcf(out_file, config)
+        else:
+            out_file = combine_variant_files(to_combine,
+                                             out_file, ref_file, config,
+                                             region=target_regions)
         if os.path.getsize(out_file) == 0:
             write_empty_vcf(out_file)
-
-        if orig_out_file.endswith(".gz"):
+        if out_file.endswith(".gz"):
             out_file = bgzip_and_index(out_file, config)
 
-        _add_reject_flag(out_file, config)
-
-def _fix_varscan_vcf(orig_file, normal_name, tumor_name, config):
-    """Fixes issues with the standard VarScan VCF output.
-
-    - Remap sample names back to those defined in the input BAM file.
-    - Convert indels into correct VCF representation.
-    """
-    tmp_file = append_stem(orig_file, "-origsample")
-
-    if not file_exists(tmp_file):
-        shutil.move(orig_file, tmp_file)
-
-        with file_transaction(config, orig_file) as tx_out_file:
-            with open(tmp_file) as in_handle:
-                with open(tx_out_file, "w") as out_handle:
-
-                    for line in in_handle:
-                        line = _clean_varscan_line(_fix_varscan_output(line, normal_name, tumor_name))
-                        if not line:
-                            continue
-                        out_handle.write(line)
-
-def _add_reject_flag(in_file, config):
-
-    """Add REJECT flag to all records that aren't flagged somatic
-    (SS=2)"""
-
-    Filter = namedtuple('Filter', ['id', 'desc'])
-    reject_filter = Filter(id='REJECT',
-                           desc='Rejected as non-SOMATIC or by quality')
-    # NOTE: PyVCF will write an uncompressed VCF
-    base, ext = utils.splitext_plus(in_file)
-    name = "rejectfix"
-    out_file = "{0}-{1}{2}".format(base, name, ".vcf")
-
-    if utils.file_exists(in_file):
-        reader = vcf.VCFReader(filename=in_file)
-        # Add info to the header of the reader
-        reader.filters["REJECT"] = reject_filter
-        with file_transaction(config, out_file) as tx_out_file:
-            with open(tx_out_file, "wb") as handle:
-                writer = vcf.VCFWriter(handle, template=reader)
-                for record in reader:
-                    if "SS" in record.INFO:
-                        # VarScan encodes it as a string
-                        # TODO: Set it as integer when cleaning
-
-                        if record.INFO["SS"] != "2":
-                            record.add_filter("REJECT")
-                    writer.write_record(record)
-
-        # Re-compress the file
-        out_file = bgzip_and_index(out_file, config)
-        move_vcf(in_file, "{0}.orig".format(in_file))
-        move_vcf(out_file, in_file)
-        with open(out_file, "w") as out_handle:
-            out_handle.write("Moved to {0}".format(in_file))
-
-
-def _fix_varscan_output(line, normal_name, tumor_name):
-    """Fix a varscan VCF line
+def fix_varscan_output(line, normal_name="", tumor_name=""):
+    """Fix a varscan VCF line.
 
     Fixes the ALT column and also fixes floating point values
     output as strings to by Floats: FREQ, SSC.
-
-    :param line: a pre-split and stripped varscan line
 
     This function was contributed by Sean Davis <sdavis2@mail.nih.gov>,
     with minor modifications by Luca Beltrame <luca.beltrame@marionegri.it>.
@@ -246,53 +138,51 @@ def _fix_varscan_output(line, normal_name, tumor_name):
         if line.startswith(tofix):
             line = line.replace('Number=1,Type=String',
                                 'Number=1,Type=Float')
-        return line + "\n"
-
+        return line
     line = line.split("\t")
 
-    mapping = {"NORMAL": normal_name, "TUMOR": tumor_name}
+    if line[0].startswith("#CHROM"):
+        if tumor_name and normal_name:
+            mapping = {"NORMAL": normal_name, "TUMOR": tumor_name}
+            base_header = line[:9]
+            old_samples = line[9:]
 
-    if(line[0].startswith("#CHROM")):
+            if len(old_samples) == 0:
+                return "\t".join(line)
 
-        base_header = line[:9]
-        old_samples = line[9:]
+            samples = [mapping[sample_name] for sample_name in old_samples]
 
-        if len(old_samples) == 0:
-            return "\t".join(line) + "\n"
-
-        samples = [mapping[sample_name] for sample_name in old_samples]
-
-        assert len(old_samples) == len(samples)
-        return "\t".join(base_header + samples) + "\n"
+            assert len(old_samples) == len(samples)
+            return "\t".join(base_header + samples)
+        else:
+            return "\t".join(line)
 
     try:
         REF, ALT = line[3:5]
     except ValueError:
-        return "\t".join(line) + "\n"
+        return "\t".join(line)
 
-    Ifreq = line[8].split(":").index("FREQ")
-    ndat = line[9].split(":")
-    tdat = line[10].split(":")
-    somatic_status = line[7].split(";")  # SS=<number>
-    # HACK: The position of the SS= changes, so we just search for it
-    somatic_status = [item for item in somatic_status
-                      if item.startswith("SS=")][0]
-    somatic_status = int(somatic_status.split("=")[1])  # Get the number
+    if len(line) > 10:
+        Ifreq = line[8].split(":").index("FREQ")
+        ndat = line[9].split(":")
+        tdat = line[10].split(":")
+        somatic_status = line[7].split(";")  # SS=<number>
+        # HACK: The position of the SS= changes, so we just search for it
+        somatic_status = [item for item in somatic_status
+                          if item.startswith("SS=")][0]
+        somatic_status = int(somatic_status.split("=")[1])  # Get the number
 
-    ndat[Ifreq] = str(float(ndat[Ifreq].rstrip("%")) / 100)
-    tdat[Ifreq] = str(float(tdat[Ifreq].rstrip("%")) / 100)
-    line[9] = ":".join(ndat)
-    line[10] = ":".join(tdat)
+        ndat[Ifreq] = str(float(ndat[Ifreq].rstrip("%")) / 100)
+        tdat[Ifreq] = str(float(tdat[Ifreq].rstrip("%")) / 100)
+        line[9] = ":".join(ndat)
+        line[10] = ":".join(tdat)
+        if somatic_status == 5:
+            # "Unknown" states are broken in current versions of VarScan
+            # so we just bail out here for now
+            return
 
     #FIXME: VarScan also produces invalid REF records (e.g. CAA/A)
     # This is not handled yet.
-
-    if somatic_status == 5:
-
-        # "Unknown" states are broken in current versions of VarScan
-        # so we just bail out here for now
-
-        return
 
     if "+" in ALT or "-" in ALT:
         if "/" not in ALT:
@@ -320,7 +210,7 @@ def _fix_varscan_output(line, normal_name, tumor_name):
 
     line[3] = REF
     line[4] = ALT
-    return "\t".join(line) + "\n"
+    return "\t".join(line)
 
 
 def _create_sample_list(in_bams, vcf_file):
@@ -344,57 +234,31 @@ def _varscan_work(align_bams, ref_file, items, target_regions, out_file):
     out_file = orig_out_file.replace(".vcf.gz", ".vcf")
 
     max_read_depth = "1000"
-    version = programs.jar_versioner("varscan", "VarScan")(config)
-    if version < "v2.3.6":
-        raise IOError("Please install version 2.3.6 or better of VarScan"
-                      " with support for multisample calling and indels"
-                      " in VCF format.")
-    varscan_jar = config_utils.get_jar("VarScan",
-                                       config_utils.get_program("varscan", config, "dir"))
     sample_list = _create_sample_list(align_bams, out_file)
     mpileup = samtools.prep_mpileup(align_bams, ref_file, config, max_read_depth,
                                     target_regions=target_regions, want_bcf=False)
     # VarScan fails to generate a header on files that start with
     # zerocoverage calls; strip these with grep, we're not going to
     # call on them
-    remove_zerocoverage = "grep -v -P '\t0\t\t$'"
-    # write a temporary mpileup file so we can check if empty
-    mpfile = "%s.mpileup" % os.path.splitext(out_file)[0]
-    with file_transaction(config, mpfile) as mpfile_tx:
-        cmd = ("{mpileup} | {remove_zerocoverage} > {mpfile_tx}")
-        do.run(cmd.format(**locals()), "mpileup for Varscan")
-    if os.path.getsize(mpfile) == 0:
-        write_empty_vcf(out_file)
-    else:
-        with tx_tmpdir(items[0]) as tmp_dir:
-            jvm_opts = _get_varscan_opts(config, tmp_dir)
-            fix_ambig = vcfutils.fix_ambiguous_cl()
-            cmd = ("cat {mpfile} "
-                   "| java {jvm_opts} -jar {varscan_jar} mpileup2cns --min-coverage 5 --p-value 0.98 "
-                   "  --vcf-sample-list {sample_list} --output-vcf --variants "
-                   "| {fix_ambig} | vcfuniqalleles > {out_file}")
-            do.run(cmd.format(**locals()), "Varscan", None,
-                   [do.file_exists(out_file)])
+    remove_zerocoverage = r"grep -v -P '\t0\t\t$'"
+    # we use ifne from moreutils to ensure we process only on files with input, skipping otherwise
+    # http://manpages.ubuntu.com/manpages/natty/man1/ifne.1.html
+    with tx_tmpdir(items[0]) as tmp_dir:
+        jvm_opts = _get_varscan_opts(config, tmp_dir)
+        fix_ambig = vcfutils.fix_ambiguous_cl()
+        py_cl = os.path.join(os.path.dirname(sys.executable), "py")
+        cmd = ("{mpileup} | {remove_zerocoverage} | "
+                "ifne varscan {jvm_opts} mpileup2cns --min-coverage 5 --p-value 0.98 "
+                "  --vcf-sample-list {sample_list} --output-vcf --variants | "
+               "{py_cl} -x 'bcbio.variation.varscan.fix_varscan_output(x)' | "
+                "{fix_ambig} | ifne vcfuniqalleles > {out_file}")
+        do.run(cmd.format(**locals()), "Varscan", None,
+                [do.file_exists(out_file)])
     os.remove(sample_list)
-    os.remove(mpfile)
     # VarScan can create completely empty files in regions without
     # variants, so we create a correctly formatted empty file
     if os.path.getsize(out_file) == 0:
         write_empty_vcf(out_file)
-    else:
-        freebayes.clean_vcf_output(out_file, _clean_varscan_line, config)
 
     if orig_out_file.endswith(".gz"):
         vcfutils.bgzip_and_index(out_file, config)
-
-
-def _clean_varscan_line(line):
-    """Avoid lines with non-GATC bases, ambiguous output bases make GATK unhappy.
-    """
-    if line and not line.startswith("#"):
-        parts = line.split("\t")
-        alleles = [x.strip() for x in parts[4].split(",")] + [parts[3].strip()]
-        for a in alleles:
-            if len(set(a) - set("GATCgatc")) > 0:
-                return None
-    return line